+++ brewtils/models.py
@@ -1,11 +1,9 @@
 # -*- coding: utf-8 -*-
 
 import copy
-from datetime import datetime
+from datetime import datetime, UTC
 from enum import Enum
-
-import pytz  # noqa # not in requirements file
-import six  # noqa # not in requirements file
+from zoneinfo import ZoneInfo
 
 from brewtils.errors import ModelError, _deprecate
 
@@ -82,10 +80,12 @@ class Events(Enum):
     GARDEN_SYNC = 30
     ENTRY_STARTED = 31
     ENTRY_STOPPED = 32
+    ENTRY_HEARTBEAT = 60
     JOB_CREATED = 33
     JOB_DELETED = 34
     JOB_PAUSED = 35
     JOB_RESUMED = 36
+    JOB_COUNTER_UPDATED = 61
     PLUGIN_LOGGER_FILE_CHANGE = 37
     RUNNER_STARTED = 38
     RUNNER_STOPPED = 39
@@ -106,7 +106,7 @@ class Events(Enum):
     REPLICATION_UPDATED = 58
     DIRECTORY_FILE_CHANGE = 59
 
-    # Next: 60
+    # Next: 62
 
 
 class Permissions(Enum):
@@ -244,6 +244,7 @@ class Instance(BaseModel):
         "STOPPING",
         "UNKNOWN",
         "AWAITING_SYSTEM",
+        "ERROR",
     }
 
     def __init__(
@@ -495,14 +496,18 @@ class StatusInfo(BaseModel):
         self.history = history or []
 
     def set_status_heartbeat(self, status, max_history=None):
-
-        self.heartbeat = datetime.utcnow()
-        self.history.append(
-            StatusHistory(status=copy.deepcopy(status), heartbeat=self.heartbeat)
-        )
+        if (
+            status != "NOT_CONFIGURED"
+            or not self.history
+            or (status == "NOT_CONFIGURED" and status != self.history[-1].status)
+        ):
+            self.heartbeat = datetime.now(UTC)
+            self.history.append(
+                StatusHistory(status=copy.deepcopy(status), heartbeat=self.heartbeat)
+            )
 
         if max_history and max_history > 0 and len(self.history) > max_history:
-            self.history = self.history[(max_history * -1) :]
+            self.history = self.history[(max_history * -1) :]  # noqa
 
     def __str__(self):
         return self.heartbeat
@@ -555,6 +560,7 @@ class File(BaseModel):
         id=None,  # noqa # shadows built-in
         owner_id=None,
         owner_type=None,
+        created_at=None,
         updated_at=None,
         file_name=None,
         file_size=None,
@@ -564,6 +570,8 @@ class File(BaseModel):
         job=None,
         request=None,
         md5_sum=None,
+        status=None,
+        root_command_type=None,
     ):
         self.id = id
         self.owner_id = owner_id
@@ -571,12 +579,15 @@ class File(BaseModel):
         self.owner = owner
         self.job = job
         self.request = request
+        self.created_at = created_at
         self.updated_at = updated_at
         self.file_name = file_name
         self.file_size = file_size
         self.chunks = chunks
         self.chunk_size = chunk_size
         self.md5_sum = md5_sum
+        self.status = status
+        self.root_command_type = root_command_type
 
     def __str__(self):
         return self.file_name
@@ -599,12 +610,20 @@ class FileChunk(BaseModel):
         offset=None,
         data=None,
         owner=None,
+        created_at=None,
+        updated_at=None,
+        status=None,
+        root_command_type=None,
     ):
         self.id = id
         self.file_id = file_id
         self.offset = offset
         self.data = data
         self.owner = owner
+        self.created_at = created_at
+        self.updated_at = updated_at
+        self.status = status
+        self.root_command_type = root_command_type
 
     def __str__(self):
         return self.data
@@ -766,6 +785,7 @@ class Request(RequestTemplate):
         output_type=None,
         status=None,
         command_type=None,
+        root_command_type=None,
         created_at=None,
         error_class=None,
         metadata=None,
@@ -805,6 +825,7 @@ class Request(RequestTemplate):
         self.requester = requester
         self.source_garden = source_garden
         self.target_garden = target_garden
+        self.root_command_type = root_command_type
 
     @classmethod
     def from_template(cls, template, **kwargs):
@@ -925,6 +946,7 @@ class System(BaseModel):
         prefix_topic=None,
         requires=None,
         requires_timeout=None,
+        garden_name=None,
     ):
         self.name = name
         self.description = description
@@ -943,15 +965,17 @@ class System(BaseModel):
         self.prefix_topic = prefix_topic
         self.requires = requires or []
         self.requires_timeout = requires_timeout
+        self.garden_name = garden_name
 
     def __str__(self):
         return "%s:%s-%s" % (self.namespace, self.name, self.version)
 
     def __repr__(self):
-        return "<System: name=%s, version=%s, namespace=%s>" % (
+        return "<System: name=%s, version=%s, namespace=%s, garden=%s>" % (
             self.name,
             self.version,
             self.namespace,
+            self.garden_name,
         )
 
     def is_newer(self, other):
@@ -1212,7 +1236,7 @@ class LoggingConfig(BaseModel):
 
         # In case no formatter is provided, we always want a default.
         formatters = {"default": {"format": self.DEFAULT_FORMAT}}
-        for formatter_name, format_str in six.iteritems(specific_formatters):
+        for formatter_name, format_str in specific_formatters.items():
             formatters[formatter_name] = {"format": format_str}
 
         return formatters
@@ -1380,7 +1404,7 @@ class Job(BaseModel):
 class DateTrigger(BaseModel):
     schema = "DateTriggerSchema"
 
-    def __init__(self, run_date=None, timezone=None):
+    def __init__(self, run_date=None, timezone="UTC"):
         self.run_date = run_date
         self.timezone = timezone
 
@@ -1396,9 +1420,10 @@ class DateTrigger(BaseModel):
 
     @property
     def scheduler_kwargs(self):
-        tz = pytz.timezone(self.timezone)
 
-        return {"timezone": tz, "run_date": tz.localize(self.run_date)}
+        tz = ZoneInfo(self.timezone.upper())
+
+        return {"timezone": tz, "run_date": self.run_date.replace(tzinfo=tz)}
 
 
 class IntervalTrigger(BaseModel):
@@ -1413,7 +1438,7 @@ class IntervalTrigger(BaseModel):
         seconds=None,
         start_date=None,
         end_date=None,
-        timezone=None,
+        timezone="UTC",
         jitter=None,
         reschedule_on_finish=None,
     ):
@@ -1455,14 +1480,16 @@ class IntervalTrigger(BaseModel):
 
     @property
     def scheduler_kwargs(self):
-        tz = pytz.timezone(self.timezone)
+        tz = ZoneInfo(self.timezone.upper())
 
         kwargs = {key: getattr(self, key) for key in self.scheduler_attributes}
         kwargs.update(
             {
                 "timezone": tz,
-                "start_date": tz.localize(self.start_date) if self.start_date else None,
-                "end_date": tz.localize(self.end_date) if self.end_date else None,
+                "start_date": (
+                    self.start_date.replace(tzinfo=tz) if self.start_date else None
+                ),
+                "end_date": self.end_date.replace(tzinfo=tz) if self.end_date else None,
             }
         )
 
@@ -1484,7 +1511,7 @@ class CronTrigger(BaseModel):
         second=None,
         start_date=None,
         end_date=None,
-        timezone=None,
+        timezone="UTC",
         jitter=None,
     ):
         self.year = year
@@ -1531,14 +1558,16 @@ class CronTrigger(BaseModel):
 
     @property
     def scheduler_kwargs(self):
-        tz = pytz.timezone(self.timezone)
+        tz = ZoneInfo(self.timezone.upper())
 
         kwargs = {key: getattr(self, key) for key in self.scheduler_attributes}
         kwargs.update(
             {
                 "timezone": tz,
-                "start_date": tz.localize(self.start_date) if self.start_date else None,
-                "end_date": tz.localize(self.end_date) if self.end_date else None,
+                "start_date": (
+                    self.start_date.replace(tzinfo=tz) if self.start_date else None
+                ),
+                "end_date": self.end_date.replace(tzinfo=tz) if self.end_date else None,
             }
         )
 
@@ -1608,25 +1637,10 @@ class FileTrigger(BaseModel):
 class Garden(BaseModel):
     schema = "GardenSchema"
 
-    GARDEN_STATUSES = {
-        "INITIALIZING",
-        "RUNNING",
-        "BLOCKED",
-        "STOPPED",
-        "NOT_CONFIGURED",
-        "CONFIGURATION_ERROR",
-        "UNREACHABLE",
-        "ERROR",
-        "UNKNOWN",
-    }
-
     def __init__(
         self,
         id=None,  # noqa # shadows built-in
         name=None,
-        status=None,
-        status_info=None,
-        namespaces=None,
         systems=None,
         connection_type=None,
         receiving_connections=None,
@@ -1641,9 +1655,6 @@ class Garden(BaseModel):
     ):
         self.id = id
         self.name = name
-        self.status = status.upper() if status else None
-        self.status_info = status_info if status_info else StatusInfo()
-        self.namespaces = namespaces or []
         self.systems = systems or []
 
         self.connection_type = connection_type
@@ -1667,11 +1678,10 @@ class Garden(BaseModel):
 
     def __repr__(self):
         return (
-            "<Garden: garden_name=%s, status=%s, version=%s, parent=%s, has_parent=%s, "
+            "<Garden: garden_name=%s, version=%s, parent=%s, has_parent=%s, "
             "connection_type=%s, receiving_connections=%s, publishing_connections=%s>"
             % (
                 self.name,
-                self.status,
                 self.version,
                 self.parent,
                 self.has_parent,